adapt core changes
