--- packages/types/src/swap_primitives.ts
@@ -1,59 +1,0 @@
-/*
- Generated by typeshare 1.13.2
-*/
-
-export interface QuoteAsset {
-	asset_id: string;
-	symbol: string;
-	decimals: number;
-}
-
-export interface ReferralAddress {
-	evm?: string;
-	solana?: string;
-	sui?: string;
-	ton?: string;
-	tron?: string;
-}
-
-export interface ReferralInfo {
-	address: ReferralAddress;
-	bps: number;
-}
-
-export interface QuoteRequest {
-	from_address: string;
-	to_address: string;
-	from_asset: QuoteAsset;
-	to_asset: QuoteAsset;
-	from_value: string;
-	referral: ReferralInfo;
-	slippage_bps: number;
-}
-
-export interface Quote {
-	quote: QuoteRequest;
-	output_value: string;
-	output_min_value: string;
-	route_data: object;
-	eta_in_seconds?: number;
-}
-
-export interface QuoteData {
-	to: string;
-	value: string;
-	data: string;
-	limit?: string;
-}
-
-export enum SwapMode {
-	ExactIn = "ExactIn",
-	ExactOut = "ExactOut",
-}
-
-export enum SwapProviderMode {
-	OnChain = "OnChain",
-	CrossChain = "CrossChain",
-	Bridge = "Bridge",
-}
-
